Fix timestamps to use the latest chrono library.
1. Refactor them into a separate file in the base.
2. Also refactor the credential stuff.
3. Remove the credential stuff from Linux.
4. Integrate the decoder and the proxy over the new Timestamp implementation.

Update versions appropriately.
