+++ adlu-base/Cargo.toml
@@ -2,7 +2,7 @@
 name = "adlu-base"
 authors = ["Daniel Brotsky <dan@clickonetwo.io>"]
 license = "AGPLv3"
-version = "1.1.0"
+version = "1.2.0"
 edition = "2021"
 
 [features]
@@ -11,7 +11,7 @@ cross-compile = []
 [dependencies]
 base64 = "0.13"
 bytes = "1"
-chrono = "=0.4.22"  # avoid deprecation warnings in 0.4.23
+chrono = "0.4"
 ctrlc = { version = "3.1", features = ["termination"] }
 eyre = "0.6"
 keyring = "1"
