+++ adlu-decoder/Cargo.toml
@@ -2,7 +2,7 @@
 name = "adlu-decoder"
 authors = ["Daniel Brotsky <dan@clickonetwo.io>"]
 license = "AGPLv3"
-version = "1.0.0"
+version = "1.0.1"
 edition = "2021"
 
 [[bin]]
@@ -11,7 +11,7 @@ name = "adlu-decoder"
 [dependencies]
 adlu-base = { path = "../adlu-base" }
 adlu-parse = { path = "../adlu-parse" }
-chrono = "=0.4.22"  # avoid deprecation warnings in 0.4.23
+chrono = "0.4"  # avoid deprecation warnings in 0.4.23
 clap = { version = "3.1.6", features = ["derive"] }
 eyre = "0.6.7"
 glob = "0.3.0"
